All Quantity Converted from int to decilam

+++ src/IICT-Store.Dtos/DistributionDtos/GetDistributionDto.cs
@@ -14,8 +14,8 @@ namespace IICT_Store.Dtos.DistributionDtos
         public int RoomNo { get; set; }
         public long ProductId { get; set; }
         public GetProductDto Product { get; set; }
-        public int Quantity { get; set; }
-        public int TotalRemainingQuantity { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal TotalRemainingQuantity { get; set; }
         public string NameOfUser { get; set; }
         public string Description { get; set; }
         public int OrderNo { get; set; }
