+++ src/IICT-Store.Dtos/ProductDtos/GetProductDto.cs
@@ -13,15 +13,15 @@ namespace IICT_Store.Dtos.ProductDtos
         public string Name { get; set; }
         public string ImageUrl { get; set; }
         public string Description { get; set; }
-        public int QuantityInStock { get; set; }
-        public int TotalQuantity { get; set; }
+        public decimal QuantityInStock { get; set; }
+        public decimal TotalQuantity { get; set; }
         public long CategoryId { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
         public GetUserDto CreatedByUser { get; set; }
         public GetUserDto UpdatedByUser { get; set; }
         public bool HasSerial { get; set; }
-        public int NotSerializedProduct { get; set; }
+        public decimal NotSerializedProduct { get; set; }
 
     }
 }
