+++ src/IICT-Store.Dtos/ProductDtos/ProductReportDto.cs
@@ -4,11 +4,11 @@
     {
         public GetProductDto Product { get; set; }
         public long ProductId { get; set; }
-        public int TotalProduct { get; set; }
-        public int TotalDistributedProduct { get; set; }
-        public int TotalDamagedProduct { get; set; }
-        public int ProductInStock { get; set; }
-        public int TotalMaintenanceProduct { get; set; }
-        public int TotalBoughtProduct { get; set; }
+        public decimal TotalProduct { get; set; }
+        public decimal TotalDistributedProduct { get; set; }
+        public decimal TotalDamagedProduct { get; set; }
+        public decimal ProductInStock { get; set; }
+        public decimal TotalMaintenanceProduct { get; set; }
+        public decimal TotalBoughtProduct { get; set; }
     }
 }
